Enabling reading bag tiddlers
Otherwise recipes that use stuff in the devstore don't work.

Originally the devstore was assumed to be used in concert with
instance config that would define store contents, so an instance
would have tiddlers and the devstore would override. It turns out
to be much more useful (from a dev standpoint) to forego store
contents until dev is stable.